Make Neff calculations available as a library
