+++ scripts/neff_gpu.py
@@ -2,158 +2,11 @@
 
 import io
 import json
-import os
 import sys
 import tarfile
 import argparse
-import time
-from string import ascii_lowercase
-from itertools import chain
-import numpy as _ # prevent "Error: mkl-service + Intel(R) MKL: MKL_THREADING_LAYER=INTEL is incompatible with libgomp.so.1 library"
 import torch
-from tqdm import tqdm
-from af22c.utils import as_handle
-from typing import NamedTuple
-from contextlib import contextmanager
-
-def loadmsa(path, stoi):
-  """load an MSA as an encoded tensor from a file-like object and a mapping dict stoi"""
-  with as_handle(path) as f:
-    lines = f.readlines()
-    assert not len(lines) & 1, f"MSA at {path} should contain an even number of lines"
-    query_id = lines[0].split()[0]
-    query_len = len([() for ch in lines[1].strip() if not ch in ascii_lowercase])
-    
-    # count header lines that are not just the query ID (1+ because first query is included in encoded msa)
-    num_seqs = 1 + sum([1 for line in lines[::2] if not line[:-1] == query_id and line.startswith(">")])
-    
-    # encode msa from lines
-    encmsa = torch.zeros((num_seqs,query_len), dtype=torch.uint8)
-    idx = 0
-    for hdr, seq in zip(lines[::2], lines[1::2]):
-      # skip query that appears multiple times in the file
-      if idx > 1 and hdr[:-1] == query_id:
-         continue
-      # encode sequence
-      encmsa[idx,:] = torch.tensor(
-        # NB: stoi comes from outside!
-        [stoi[ch] for ch in seq[:-1] if not ch in ascii_lowercase],
-        dtype=torch.uint8,
-      )
-      idx += 1
-    return encmsa
-
-class EncMsa(NamedTuple):
-  """Store an encoded MSA with vocabulary and means of encoding"""
-  vocab: list[str] # list of characters used in the MSA, including gaps
-  stoi: dict # mapping: "residue character" -> int
-  itos: dict # mapping: int -> "residue character"
-  data: torch.Tensor # (N, L) matrix, where N is the number of sequences and L is the length of the query sequence
-  gaptok: int # integer version of the gap token
-  
-  @property
-  def shape(self):
-    return self.data.shape
-  
-  @property
-  def device(self):
-    return self.data.device
-  
-  def to(self, device):
-    """mimic torch behavior and move data to `device`. (return a new tuple since `self` is immutable.)"""
-    params = self._asdict()
-    params["data"] = self.data.to(device)
-    return EncMsa(**params)
-
-@contextmanager
-def as_encmsa(thing):
-  """convert "any"`thing` (that is convertible to MSA) to an MSA"""
-  if isinstance(thing, EncMsa): # invariant against `EncMsa`s
-    yield thing
-  else:
-    # vocabulary and conversion
-    vocab = '-ACDEFGHIKLMNPQRSTVWXY' # TODO: do we want to infer vocab?
-    stoi = {c:i for i, c in enumerate(vocab)}
-    itos = {i:c for c, i in stoi.items()}
-    gaptok = stoi['-']
-    
-    # try to convert the MSA if not done yet
-    if isinstance(thing, torch.Tensor):
-      encmsa = thing
-    elif isinstance(thing, str) or isinstance(thing, io.TextIOWrapper):
-      encmsa = loadmsa(thing, stoi)
-    else:
-      raise ValueError(f"unable to interpret type of input as msa ({type(thing)=})")
-    yield EncMsa(vocab, stoi, itos, encmsa, gaptok)
-
-def pwseq(msa, device=None, batch_size=2**12, verbose=True, **kwargs):
-  """return pairwise sequence identity calculated with pytorch"""
-  with as_encmsa(msa) as encmsa:
-    num_seqs,query_len = encmsa.shape
-    
-    # calculate all pairs for which pairwise sequence identities need to be calculated
-    pairs = torch.triu_indices(*(num_seqs,)*2, 1, device=device).T
-    
-    # each batch should yield a matrix with batch_size elements
-    num_batches = (len(pairs) + batch_size - 1) // batch_size
-    num_full_batches = len(pairs) // batch_size
-    
-    batch_pairs = pairs[:-(len(pairs)%batch_size)]
-    rest_pairs = pairs[-(len(pairs)%batch_size):]
-    
-    # put pairs into batches
-    batches = batch_pairs.view(max(num_full_batches,1), -1, 2)
-    if num_batches != num_full_batches:
-      batches = chain(batches, [rest_pairs])
-      
-    # one batch contains batch_size many pairs, which yields batch_size many similarity scores because the 
-    # similarity matrix is symmetric.
-    bpwseq = torch.eye(num_seqs, device=device) # matrix containing similarity scores for two sequences
-    
-    for batch_pairs in (
-      tqdm(batches, total=num_batches, desc="running batches")
-      if verbose else
-      batches
-    ):
-      # extract sequences in batch
-      batch_seqs = encmsa.data[batch_pairs]
-
-      # calculate pairwise distances 
-      batch_pwdists = torch.sum(batch_seqs[:,0,:] != batch_seqs[:,1,:], axis=-1)
-      batch_pwseq = 1 - batch_pwdists / query_len
-      
-      # put at right location in result matrix (and make symmetric)
-      bpwseq[batch_pairs[:,0],batch_pairs[:,1]] = batch_pwseq
-      bpwseq[batch_pairs[:,1],batch_pairs[:,0]] = batch_pwseq
-    
-    return bpwseq
-  
-def gapcount(msa, weights=None, nongap=False, **kwargs):
-  """
-  calculate number of gaps in each msa column. returns a vector of length query_len.
-  weights can be specified to weight each sequence in the msa column.
-  if nongap=False (default), gaps will be counted; otherwise non-gaps will be counted.
-  """
-  with as_encmsa(msa) as encmsa:
-    gapindicator = encmsa.data != encmsa.gaptok if nongap else encmsa.data == encmsa.gaptok
-    if weights is not None:
-      return weights @ gapindicator.float()
-    return torch.sum(gapindicator, dim=0)
-
-def neff(msa, pwseqfn=pwseq, seqid_thres=0.8, **kwargs):
-  """calculate neff scores for an encoded msa."""
-  device = kwargs.get("device")
-  
-  with as_encmsa(msa) as encmsa:
-    # put msa on desired device, if it is not there already
-    if encmsa.device != device:
-      encmsa = encmsa.to(device)
-    
-    num_seqs, query_len = encmsa.shape
-    eq = pwseqfn(encmsa, **kwargs)
-    # calculate neff weights (dim can be 0 or 1, does not matter because pwseq is symmetric)
-    neffweights = 1 / torch.sum(eq >= seqid_thres, dim=0)
-    return gapcount(encmsa, weights=neffweights, nongap=True, **kwargs)
+from af22c.neff import neff, gapcount
 
 def open_a3m(archive_path, a3m_path):
   """
